Directional lights now supported
