+++ game-engine/include/game-engine/Core/Modules/Graphics/Graphics.h
@@ -19,7 +19,6 @@ class Texture;
 class Graphics : public CoreModule
 {
 public:
-    static const unsigned int MAX_LIGHTS;
     
 	static Graphics& getInstance()
 	{
